[CHORE] StarRate 컴포넌트 setRate props 추가 #
